feat(compiler): add support for object literal types in the Diez compiler (#310)
Object literal types may be defined in any context with an expression like this:

```
const palette = {
  blue: Color.rgba(...),
};
```

When we encounter this "expando" type, provided that the pascal-case type name derived from the variable assignment in which it is declared is globally unique (e.g. `palette` -> `Palette`), the Diez compiler can use this as a true singleton.

Class declarations are still supported. `target` and `stdlib` fixtures are updated to use the new syntax, and a small slice of `poodle-surf` was upgraded to the new singleton syntax to demonstrate backward compatibility.